% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{mixedmodeling_deps}
\alias{mixedmodeling_deps}
\title{List all mixedmodeling dependencies}
\usage{
mixedmodeling_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
mixedmodeling packages.}
}
\description{
List all mixedmodeling dependencies
}
