% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{mixedmodeling_packages}
\alias{mixedmodeling_packages}
\title{List all packages in the mixedmodeling}
\usage{
mixedmodeling_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include mixedmodeling in the list?}
}
\description{
List all packages in the mixedmodeling
}
\examples{
mixedmodeling_packages()
}
