% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{mixedmodeling_update}
\alias{mixedmodeling_update}
\title{Update mixedmodeling packages}
\usage{
mixedmodeling_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
mixedmodeling packages.}
}
\description{
This will check to see if all mixedmodeling packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
mixedmodeling_update()
}
}
